# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Skybull Snapser CloudeCode",
    "description": "Skybull snapser cloud code",
    "contact": {},
    "version": "0.0.1"
  },
...
  "servers": [
    {
      "url": "//localhost:8080/"
    }
  ],
  "paths": {
...
    "/v1/byosnap-skybull-cloudcode/create-campaign-game": {
      "post": {
        "summary": "Call when start a campaign game",
        "description": "Create campaign game and subject energy",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "Authorization Token",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "sessionToken",
                  "userId"
                ],
                "properties": {
                  "userId": {
                    "type": "string",
                    "description": "userId when they login"
                  },
                  "sessionToken": {
                    "type": "string",
                    "description": "session token"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/v1/byosnap-skybull-cloudcode/create-pvp-game": {
      "post": {
        "summary": "Level up equipment",
        "description": "Level up equipment",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "Authorization Token",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "matchId",
                  "opponentID",
                  "sessionToken",
                  "userId"
                ],
                "properties": {
                  "userId": {
                    "type": "string",
                    "description": "userId when they login"
                  },
                  "sessionToken": {
                    "type": "string",
                    "description": "session token"
                  },
                  "matchId": {
                    "type": "string",
                    "description": "id of the pvp game"
                  },
                  "opponentID": {
                    "type": "string",
                    "description": "id of opponent"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/byosnap-skybull-cloudcode/win-campaign": {
      "post": {
        "summary": "Call when win a campaign",
        "description": "Call when win a campaign",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "Authorization Token",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "sessionToken",
                  "stagePower",
                  "star",
                  "userId"
                ],
                "properties": {
                  "userId": {
                    "type": "string",
                    "description": "userId when they login with snapser"
                  },
                  "sessionToken": {
                    "type": "string",
                    "description": "session token for login version"
                  },
                  "stagePower": {
                    "type": "string",
                    "description": "power stage player win"
                  },
                  "star": {
                    "type": "string",
                    "description": "the start they reach"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/v1/byosnap-skybull-cloudcode/win-pvp": {
      "post": {
        "summary": "Call when player win a pvp game",
        "description": "Call when player win a pvp game",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "Authorization Token",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "opponentID",
                  "sessionToken",
                  "userId"
                ],
                "properties": {
                  "userId": {
                    "type": "string",
                    "description": "userId when they login"
                  },
                  "sessionToken": {
                    "type": "string",
                    "description": "session token"
                  },
                  "opponentID": {
                    "type": "string",
                    "description": "opponent snapser id "
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BasicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}